refactor: changes files packages and add api-docs route
