+++ src/routes/health-check/healthCheckRoute.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import healthCheckController from "../health-check-controller/healthCheck.controller";
+import healthCheckController from "../../monitoring/health-check/health-check-controller/healthCheck.controller";
 
 const healthCheckRoute = Router();
