+++ src/routes/index.ts
@@ -1,10 +1,12 @@
 import { Router } from "express";
-import sensorRoutes from "./v1/Sensors/sensorRoute";
-import healthCheckRoute from "./health-check/health-check-route/healthCheckRoute";
 import corsMiddleware from "../middleware/corsMiddleware";
+import sensorRoutes from "./v1/Sensors/sensorRoute";
+import healthCheckRoute from "./health-check/healthCheckRoute";
+import DocsRoute from "./docs/docsRoute";
 
 const router = Router();
 
+router.use("/api-docs", corsMiddleware.execute, DocsRoute);
 router.use("/health-check", corsMiddleware.execute, healthCheckRoute);
 router.use("/sensors", corsMiddleware.execute, sensorRoutes);
